+++ src/binlog/trunk/trunk_binlog.h
@@ -29,7 +29,7 @@ extern "C" {
     int trunk_sf_binlog_get_current_write_index();
 
     int trunk_binlog_write(const char op_type, const int path_index,
-            const DATrunkIdInfo *id_info, const int64_t file_size);
+            const DATrunkIdInfo *id_info, const uint32_t file_size);
 
 #ifdef __cplusplus
 }
